requirements.txt
requests
python-dateutil
tz